Combining and reorganizing constants
	modified:   src/main/java/frc/robot/Constants.java
	modified:   src/main/java/frc/robot/RobotContainer.java
	modified:   src/main/java/frc/robot/commands/ChoreoAutoController.java
	modified:   src/main/java/frc/robot/subsystems/drive/Drive.java
	modified:   src/main/java/frc/robot/subsystems/flywheel_example/FlywheelIOSpark.java
	modified:   src/main/java/frc/robot/subsystems/flywheel_example/FlywheelIOTalonFX.java
	modified:   src/main/java/frc/robot/subsystems/vision/Vision.java
	modified:   src/main/java/frc/robot/subsystems/vision/VisionIOPhotonVisionSim.java
	modified:   src/main/java/frc/robot/util/PowerMonitoring.java
