+++ src/main/java/frc/robot/subsystems/vision/VisionIOPhotonVisionSim.java
@@ -15,10 +15,9 @@
 
 package frc.robot.subsystems.vision;
 
-import static frc.robot.Constants.VisionConstants2.aprilTagLayout;
-
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Transform3d;
+import frc.robot.Constants.AprilTagConstants;
 import java.util.function.Supplier;
 import org.photonvision.simulation.PhotonCameraSim;
 import org.photonvision.simulation.SimCameraProperties;
@@ -45,7 +44,7 @@ public class VisionIOPhotonVisionSim extends VisionIOPhotonVision {
     // Initialize vision sim
     if (visionSim == null) {
       visionSim = new VisionSystemSim("main");
-      visionSim.addAprilTags(aprilTagLayout);
+      visionSim.addAprilTags(AprilTagConstants.aprilTagLayout);
     }
 
     // Add sim camera
